rename gh_strings
